Added new custom JSON filters for the search API and improved code quality in the crawler

+++ pkg/crawler/types.go
@@ -39,7 +39,7 @@ type Pars struct {
 	SelIdx  int
 	RE      *rules.RuleEngine
 	Sources *[]cdb.Source
-	Index   int
+	Index   uint64
 	Status  *Status
 }
